Moved grouping of parameters to html output
For #25
