+++ lib/htmlWriter.js
@@ -4,6 +4,7 @@ const fse = require('fs-extra')
 const glob = require('glob')
 const cheerio = require('cheerio')
 const mu = require('mustache')
+const unique = require('array-unique')
 
 const js = require('../lib/jsonquery')
 const ut = require('../lib/util')
@@ -275,6 +276,13 @@ function toHtml (component) {
 
   // Add html elements to simplify output generation
   var comCop = addHtmlElements(component)
+  // Group all the parameters by tabs and groups
+  if (comCop.public && comCop.public.parameters) {
+    comCop.public.parameters = groupParameters(comCop.public.parameters)
+  }
+  if (comCop.protected && comCop.protected.parameters) {
+    comCop.protected.parameters = groupParameters(comCop.protected.parameters)
+  }
 
   // Not all blocks have an html section
   if (comCop.info) {
@@ -312,6 +320,55 @@ function toHtml (component) {
     comCop.info = comCop.info.replace('</html>', '').replace('<html>', '')
   }
   return mu.render(template, comCop).trim()
+}
+
+/** Return a list with the elements of `listOfPars` that are
+  * in the tab `tab` and in the group `group`.
+  *
+  *@param listOfPars A list of parameters.
+  *@param tab A string with the tab name.
+  *@param group A string with the group name.
+  *@return A list with the parameters filtered by `tab` and `group`.
+  *        If there is no match, then `null` is returned.
+  */
+var getParametersByTabAndGroup = function (listOfPars, tab, group) {
+  return listOfPars.filter(par => {
+    const thisTab = js.getProperty(['annotation', 'dialog', 'tab'], par)
+    if (thisTab !== tab) {
+      return null
+    } else {
+      const thisGro = js.getProperty(['annotation', 'dialog', 'group'], par)
+      return (thisGro === group)
+    }
+  })
+}
+
+/** Return a json structure of the parameters, grouped by
+  * tab and group as received from their annotation.
+  *
+  *@param listOfPars A list of parameters.
+  *@return A json structure with the parameters grouped by `tab` and `group`.
+  */
+function groupParameters (listOfPars) {
+  var tabs = listOfPars.map(par => { return js.getProperty(['annotation', 'dialog', 'tab'], par) })
+
+  // Remove dublicates and null
+  const tabsUni = unique(tabs).filter(ele => ele !== null)
+
+  const tabEntries = tabsUni.map(tab => {
+    // Within each tab, group the parameters by their group
+    const parOnTab = listOfPars.filter(par => js.getProperty(['annotation', 'dialog', 'tab'], par) === tab)
+    var groups = parOnTab.map(par => { return js.getProperty(['annotation', 'dialog', 'group'], par) })
+    // Remove dublicats and null
+    const groUni = unique(groups).filter(ele => ele !== null)
+    var parsInGroup = []
+    groUni.forEach(gro => {
+      parsInGroup.push({ 'name': gro, 'parameters': getParametersByTabAndGroup(listOfPars, tab, gro) })
+    })
+    return { 'name': tab, 'groups': parsInGroup }
+  })
+
+  return {'tabs': tabEntries}
 }
 
 function removeHtmlTag (string, tag) {
@@ -401,15 +458,11 @@ function addHtmlToConnectors (path, components) {
 
 function addHtmlToParameters (path, components) {
   const prop = js.getProperty(path, components)
-  if (prop) {
+  if (prop !== null) {
     // Iterate over the tabs and groups
-    prop.tabs.forEach(tab => {
-      tab.groups.forEach(gro => {
-        gro.parameters.forEach(par => {
-          setParameterHTMLAttributes(par)
-          setHtmlToParameterModification(par)
-        })
-      })
+    prop.forEach(par => {
+      setParameterHTMLAttributes(par)
+      setHtmlToParameterModification(par)
     })
   }
 }
