+++ lib/jsonquery.js
@@ -1,5 +1,4 @@
 // const fs = require('fs')
-const unique = require('array-unique')
 var logger = require('winston')
 
 /** Get the JSON property `p` from the json data `o`
@@ -9,54 +8,6 @@ const getProperty = (p, o) =>
   p.reduce((xs, x) =>
     (xs && xs[x]) ? xs[x] : null, o)
 
-/** Return a list with the elements of `listOfPars` that are
-  * in the tab `tab` and in the group `group`.
-  *
-  *@param listOfPars A list of parameters.
-  *@param tab A string with the tab name.
-  *@param group A string with the group name.
-  *@return A list with the parameters filtered by `tab` and `group`.
-  *        If there is no match, then `null` is returned.
-  */
-var getParametersByTabAndGroup = function (listOfPars, tab, group) {
-  return listOfPars.filter(par => {
-    const thisTab = getProperty(['annotation', 'dialog', 'tab'], par)
-    if (thisTab !== tab) {
-      return null
-    } else {
-      const thisGro = getProperty(['annotation', 'dialog', 'group'], par)
-      return (thisGro === group)
-    }
-  })
-}
-
-/** Return a json structure of the parameters, grouped by
-  * tab and group as received from their annotation.
-  *
-  *@param listOfPars A list of parameters.
-  *@return A json structure with the parameters grouped by `tab` and `group`.
-  */
-function groupParameters (listOfPars) {
-  var tabs = listOfPars.map(par => { return getProperty(['annotation', 'dialog', 'tab'], par) })
-
-  // Remove dublicates and null
-  const tabsUni = unique(tabs).filter(ele => ele !== null)
-
-  const tabEntries = tabsUni.map(tab => {
-    // Within each tab, group the parameters by their group
-    const parOnTab = listOfPars.filter(par => getProperty(['annotation', 'dialog', 'tab'], par) === tab)
-    var groups = parOnTab.map(par => { return getProperty(['annotation', 'dialog', 'group'], par) })
-    // Remove dublicats and null
-    const groUni = unique(groups).filter(ele => ele !== null)
-    var parsInGroup = []
-    groUni.forEach(gro => {
-      parsInGroup.push({ 'name': gro, 'parameters': getParametersByTabAndGroup(listOfPars, tab, gro) })
-    })
-    return { 'name': tab, 'groups': parsInGroup }
-  })
-
-  return {'tabs': tabEntries}
-}
 /** Get the simplified json representation for the model
   */
 function simplifyModelicaJSON (model) {
@@ -116,9 +67,6 @@ function simplifyModelicaJSON (model) {
   if (connections) {
     addConnectorTargets(connections, allModels)
   }
-  // Group all the parameters by tabs and groups
-  const pubParGrouped = groupParameters(pubParameters)
-  const proParGrouped = groupParameters(proParameters)
 
   var data = {
     'modelicaFile': model.modelicaFile,
@@ -126,13 +74,13 @@ function simplifyModelicaJSON (model) {
     'topClassName': className,
     'comment': getComment(model).replace(/"/g, ''),
     'public': {
-      'parameters': pubParGrouped,
+      'parameters': pubParameters,
       'models': pubModels,
       'inputs': pubInputs,
       'outputs': pubOutputs
     },
     'protected': {
-      'parameters': proParGrouped,
+      'parameters': proParameters,
       'models': proModels,
       'inputs': proInputs,
       'outputs': proOutputs
