+++ test/FromModelica/Parameter1-simplified.json
@@ -5,32 +5,20 @@
     "topClassName": "FromModelica.Parameter1",
     "comment": "Some class comment",
     "public": {
-      "parameters": {
-        "tabs": [
-          {
-            "name": "General",
-            "groups": [
-              {
-                "name": "Parameters",
-                "parameters": [
-                  {
-                    "name": "kP",
-                    "value": "1",
-                    "comment": "Some comment",
-                    "annotation": {
-                      "dialog": {
-                        "tab": "General",
-                        "group": "Parameters"
-                      }
-                    },
-                    "className": "Real"
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      }
+      "parameters": [
+        {
+          "name": "kP",
+          "value": "1",
+          "comment": "Some comment",
+          "annotation": {
+            "dialog": {
+              "tab": "General",
+              "group": "Parameters"
+            }
+          },
+          "className": "Real"
+        }
+      ]
     },
     "info": null
   }
