+++ test/FromModelica/ParameterWithInfo-simplified.json
@@ -5,32 +5,20 @@
     "topClassName": "FromModelica.ParameterWithInfo",
     "comment": "Some class comment",
     "public": {
-      "parameters": {
-        "tabs": [
-          {
-            "name": "General",
-            "groups": [
-              {
-                "name": "Parameters",
-                "parameters": [
-                  {
-                    "name": "kP",
-                    "value": "1",
-                    "comment": "Some comment",
-                    "annotation": {
-                      "dialog": {
-                        "tab": "General",
-                        "group": "Parameters"
-                      }
-                    },
-                    "className": "Real"
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      }
+      "parameters": [
+        {
+          "name": "kP",
+          "value": "1",
+          "comment": "Some comment",
+          "annotation": {
+            "dialog": {
+              "tab": "General",
+              "group": "Parameters"
+            }
+          },
+          "className": "Real"
+        }
+      ]
     },
     "info": "<html>\n<p>\nThis is the info section.\n</p>\n</html>"
   }
